Add annotations form and introduce the picker modal
Annotations can be added. Selecting the original data source and sample
frequencies is done using the new picker modal.

Introduce the picker modal, a list view of resources in a modal popup
window from which the user can select a resource. The picker modal is
opened by clicking on the associated form field. In addition to the add
annotation form, the picker modal is used to select a sample in the
variant and region lookup forms.