Use Promises (bluebird) in variants listing
This is a temporary implementation with some promisification wrappers
for our API methods. The code is already much cleaner than the old
version with nested async callbacks. I think eventually we should
rewrite the API methods to return Promises directly.

This is in anticipation of yet more nesting when we start to query
mygene.info for transcript annotation.